chapter_4_backend_events/backend/pyproject.toml
[project]
name = "voice-escape-room-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "python-dotenv",
    "websockets",
]
